Remove unnecessary clippy lints and fix them properly

+++ lib/compiler-singlepass/src/machine_arm64.rs
@@ -168,7 +168,6 @@ impl MachineARM64 {
         }
     }
 
-    #[allow(clippy::unnecessary_cast)]
     fn location_to_reg(
         &mut self,
         sz: Size,
@@ -235,8 +234,7 @@ impl MachineARM64 {
                         temps.push(tmp);
                         tmp
                     };
-                    self.assembler
-                        .emit_mov_imm(Location::GPR(tmp), val as u64)?;
+                    self.assembler.emit_mov_imm(Location::GPR(tmp), val)?;
                     Ok(Location::GPR(tmp))
                 }
             }
@@ -321,7 +319,6 @@ impl MachineARM64 {
             _ => codegen_error!("singlepass can't emit location_to_reg {:?} {:?}", sz, src),
         }
     }
-    #[allow(clippy::unnecessary_cast)]
     fn location_to_neon(
         &mut self,
         sz: Size,
@@ -391,8 +388,7 @@ impl MachineARM64 {
                         CompileError::Codegen("singlepass cannot acquire temp simd".to_owned())
                     })?;
                     temps.push(tmp);
-                    self.assembler
-                        .emit_mov_imm(Location::GPR(gpr), val as u64)?;
+                    self.assembler.emit_mov_imm(Location::GPR(gpr), val)?;
                     self.assembler
                         .emit_mov(sz, Location::GPR(gpr), Location::SIMD(tmp))?;
                     self.release_gpr(gpr);
